Bridge/requirements.txt
PySide6
python-socketio[client]
requests
